Adds MissingStringLocalization and MissingImageLocalization errors.
